+++ arbiter-core/Cargo.toml
@@ -32,3 +32,4 @@ tokio = { version = "1.28.1", features = ["macros", "full"] }
 
 [dev-dependencies]
 hex = { version = "0.4.3", default-features = false }
+
