+++ arbiter-core/src/tests/bindings.rs
@@ -1,0 +1,12 @@
+use super::*;
+
+// TODO: Hit all the contract bindings.
+#[test]
+fn token_mint() -> Result<()> {
+    todo!()
+}
+
+#[test]
+fn arbiter_math() -> Result<()> {
+    todo!()
+}
