+++ arbiter-core/src/tests/interaction.rs
@@ -1,183 +1,8 @@
-#![allow(missing_docs)]
-use std::{str::FromStr, os::unix::thread};
-
-use anyhow::{Ok, Result};
-use ethers::{
-    prelude::{EthLogDecode, Middleware, StreamExt},
-    types::{Address, Filter, ValueOrArray, U64},
-};
-
-use crate::{
-    agent::{tests::TEST_AGENT_NAME, *},
-    bindings::arbiter_token::*,
-    environment::{tests::TEST_ENV_LABEL, *},
-    manager::*,
-    math::*,
-    middleware::*,
-};
-
-pub const TEST_ARG_NAME: &str = "ArbiterToken";
-pub const TEST_ARG_SYMBOL: &str = "ARBT";
-pub const TEST_ARG_DECIMALS: u8 = 18;
-pub const TEST_MINT_AMOUNT: u128 = 69;
-pub const TEST_MINT_TO: &str = "0xf7e93cc543d97af6632c9b8864417379dba4bf15";
-pub const TEST_APPROVAL_AMOUNT: u128 = 420;
-
-//TODO: Send a tx before and after pausing the environment.
-
-#[test]
-fn add_environment() {
-    let mut manager = SimulationManager::new();
-    manager.add_environment(TEST_ENV_LABEL, 1.0, 1).unwrap();
-    assert!(manager
-        .environments
-        .contains_key(&TEST_ENV_LABEL.to_string()));
-    assert_eq!(
-        manager
-            .environments
-            .get(&TEST_ENV_LABEL.to_string())
-            .unwrap()
-            .state
-            .load(std::sync::atomic::Ordering::Relaxed),
-        State::Initialization
-    );
-}
-
-#[test]
-fn run_environment() {
-    let mut manager = SimulationManager::new();
-    manager.add_environment(TEST_ENV_LABEL, 1.0, 1).unwrap();
-    manager.start_environment(TEST_ENV_LABEL).unwrap();
-    assert_eq!(
-        manager
-            .environments
-            .get(&TEST_ENV_LABEL.to_string())
-            .unwrap()
-            .state
-            .load(std::sync::atomic::Ordering::Relaxed),
-        State::Running
-    );
-}
-
-#[test]
-fn pause_environment() {
-    let mut manager = SimulationManager::new();
-    manager.add_environment(TEST_ENV_LABEL, 1.0, 1).unwrap();
-    manager.start_environment(TEST_ENV_LABEL).unwrap();
-    manager.pause_environment(TEST_ENV_LABEL).unwrap();
-    std::thread::sleep(std::time::Duration::from_millis(100));
-    assert_eq!(
-        manager
-            .environments
-            .get(&TEST_ENV_LABEL.to_string())
-            .unwrap()
-            .state
-            .load(std::sync::atomic::Ordering::Relaxed),
-        State::Paused
-    );
-    manager.start_environment(TEST_ENV_LABEL).unwrap();
-    std::thread::sleep(std::time::Duration::from_millis(100));
-    assert_eq!(
-        manager
-            .environments
-            .get(&TEST_ENV_LABEL.to_string())
-            .unwrap()
-            .state
-            .load(std::sync::atomic::Ordering::Relaxed),
-        State::Running
-    );
-}
-
-#[test]
-fn stop_environment() {
-    let mut manager = SimulationManager::new();
-    manager.add_environment(TEST_ENV_LABEL, 1.0, 1).unwrap();
-    manager.start_environment(TEST_ENV_LABEL).unwrap();
-    manager.stop_environment(TEST_ENV_LABEL).unwrap();
-    assert_eq!(
-        manager
-            .environments
-            .get(&TEST_ENV_LABEL.to_string())
-            .unwrap()
-            .state
-            .load(std::sync::atomic::Ordering::Relaxed),
-        State::Stopped
-    );
-}
-
-// TODO: Hit all the contract bindings.
-#[test]
-fn token_mint() -> Result<()> {
-    todo!()
-}
-
-#[test]
-fn arbiter_math() -> Result<()> {
-    todo!()
-}
-
-#[test]
-fn attach_agent() {
-    let environment = &mut Environment::new(TEST_ENV_LABEL, 1.0, 1);
-    let agent = Agent::new(TEST_AGENT_NAME);
-    agent.attach_to_environment(environment);
-    assert_eq!(environment.agents[0].name, TEST_AGENT_NAME);
-}
-
-#[test]
-fn simulation_agent_wallet() {
-    let environment = &mut Environment::new(TEST_ENV_LABEL, 1.0, 1);
-    let agent = Agent::new(TEST_AGENT_NAME);
-    agent.attach_to_environment(environment);
-    assert_eq!(
-        environment.agents[0].client.default_sender().unwrap(),
-        Address::from_str("0x09e12ce98726acd515b68f87f49dc2e5558f6a72").unwrap()
-    );
-}
-
-#[test]
-fn multiple_agent_addresses() {
-    let environment = &mut Environment::new(TEST_ENV_LABEL, 1.0, 1);
-    let agent = Agent::new(TEST_AGENT_NAME);
-    agent.attach_to_environment(environment);
-    let agent2 = Agent::new(format!("new_{}", TEST_AGENT_NAME));
-    agent2.attach_to_environment(environment);
-    assert_ne!(
-        environment.agents[0].client.default_sender(),
-        environment.agents[1].client.default_sender()
-    );
-}
-
-// TODO: Test to see that we prvent agents with the same name from being added.
-#[test]
-fn agent_name_collision() {
-    todo!();
-}
-
-async fn deploy_and_start() -> Result<(ArbiterToken<RevmMiddleware>, Environment)> {
-    let mut environment = Environment::new(TEST_ENV_LABEL, 1.0, 1);
-    let agent = Agent::new(TEST_AGENT_NAME);
-    agent.attach_to_environment(&mut environment);
-    environment.run();
-    Ok((
-        ArbiterToken::deploy(
-            environment.agents[0].client.clone(),
-            (
-                TEST_ARG_NAME.to_string(),
-                TEST_ARG_SYMBOL.to_string(),
-                TEST_ARG_DECIMALS,
-            ),
-        )?
-        .send()
-        .await
-        .unwrap(),
-        environment,
-    ))
-}
+use super::*;
 
 #[tokio::test]
 async fn test_deploy() -> Result<()> {
-    let (arbiter_token, environment) = deploy_and_start().await?;
+    let (arbiter_token, _environment) = deploy_and_start().await?;
     println!("{:?}", arbiter_token);
     assert_eq!(
         arbiter_token.address(),
@@ -372,7 +197,7 @@ async fn filter_topics() -> Result<()> {
 
     // Use tokio::time::timeout to await the approval_watcher for a specific duration
     // The timeout is needed because the approval_watcher is a stream that will never end when the test is passing
-    let timeout_duration = tokio::time::Duration::from_secs(1); // Adjust the duration as needed
+    let timeout_duration = tokio::time::Duration::from_secs(5); // Adjust the duration as needed
     let timeout = tokio::time::timeout(timeout_duration, approval_watcher.next());
     match timeout.await {
         Result::Ok(Some(_)) => {
